requirements.txt
numpy==1.23.5
pandas==2.2.1
ipykernel==6.29.3
ipython==8.7.0
jupyter-client==7.4.7
jupyter-core==5.1.0
matplotlib==3.6.2
python-dotenv==1.0.1
langchain==0.1.13
langchain_openai==0.1.1
openpyxl==3.1.2 
python-docx==1.1.0
faiss-cpu==1.8.0
pdfminer.six==20231228
streamlit==1.32.2